+++ build_dist.sh
@@ -4,7 +4,7 @@ set -e
 
 cd "$(dirname "$0")"
 
-rm -rf build/ dist/ quench.egg-info/
+rm -rf build/ dist/ pyquench.egg-info/
 
 python3 -m build
 python3 -m twine check dist/*